Composite handling moved to its own namespace
