+++ src/Npgsql/TypeHandlers/CompositeHandles/IDynamicCompositeTypeHandlerFactory.cs
@@ -1,0 +1,6 @@
+namespace Npgsql.TypeHandlers.CompositeHandlers
+{
+#pragma warning disable CA1040    // Avoid empty interfaces
+    interface IDynamicCompositeTypeHandlerFactory { }
+#pragma warning restore CA1040    // Avoid empty interfaces
+}
