+++ src/Npgsql/TypeHandlers/CompositeHandles/IMappedCompositeHandler.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Npgsql.TypeHandlers.CompositeHandlers
+{
+    interface IMappedCompositeHandler
+    {
+        /// <summary>
+        /// The CLR type mapped to the PostgreSQL composite type.
+        /// </summary>
+        Type CompositeType { get; }
+    }
+}
