+++ src/Npgsql/TypeHandlers/CompositeHandles/MappedCompositeTypeHandlerFactory.cs
@@ -1,0 +1,23 @@
+using System;
+using Npgsql.PostgresTypes;
+using Npgsql.TypeHandling;
+
+namespace Npgsql.TypeHandlers.CompositeHandlers
+{
+    class MappedCompositeTypeHandlerFactory<T> : NpgsqlTypeHandlerFactory<T>, IMappedCompositeTypeHandlerFactory
+        where T : new()
+    {
+        public INpgsqlNameTranslator NameTranslator { get; }
+
+        internal MappedCompositeTypeHandlerFactory(INpgsqlNameTranslator nameTranslator)
+        {
+            NameTranslator = nameTranslator;
+        }
+
+        internal override NpgsqlTypeHandler Create(PostgresType pgType, NpgsqlConnection conn)
+            => new MappedCompositeHandler<T>(NameTranslator, pgType, conn);
+
+        protected override NpgsqlTypeHandler<T> Create(NpgsqlConnection conn)
+            => throw new InvalidOperationException($"Expect {nameof(PostgresType)}");
+    }
+}
